+++ packages/auto-scroll/src/internal-types.ts
@@ -42,22 +42,22 @@ export type InternalConfig = {
   maxPixelScrollPerSecond: number;
   timeDampeningDurationMs: number;
   maxMainAxisHitboxSize: number;
-  allowedAxis: AllowedAxis;
 };
 
 export type PublicConfig = Partial<{
   maxScrollSpeed: 'standard' | 'fast';
-  allowedAxis: AllowedAxis;
 }>;
 
 export type ElementAutoScrollArgs<DragType extends AllDragTypes> = {
   element: Element;
   canScroll?: (args: ElementGetFeedbackArgs<DragType>) => boolean;
+  getAllowedAxis?: (args: ElementGetFeedbackArgs<DragType>) => AllowedAxis;
   getConfiguration?: (args: ElementGetFeedbackArgs<DragType>) => PublicConfig;
 };
 
 export type WindowAutoScrollArgs<DragType extends AllDragTypes> = {
   canScroll?: (args: WindowGetFeedbackArgs<DragType>) => boolean;
+  getAllowedAxis?: (args: WindowGetFeedbackArgs<DragType>) => AllowedAxis;
   getConfiguration?: (args: WindowGetFeedbackArgs<DragType>) => PublicConfig;
 };